python/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "filetimefixer"
version = "0.2.0"
description = "Fix image file timestamps and EXIF from filename/time for correct sort order on Windows, Mac, Android, iPhone."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "FileTimeFixer" }]
keywords = ["exif", "photo", "timestamp", "rename"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "Pillow>=9.0",
    "piexif>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "PyYAML>=6.0",
]

[project.scripts]
filetimefixer = "filetimefixer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["filetimefixer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["filetimefixer"]
branch = true

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:", "raise NotImplementedError"]
